test/regression/reftest/web-platform-tests/dom/events/Event-dispatch-omitted-capture.html

<!DOCTYPE html>
<meta charset=utf-8>
<!--EDIT-->
<!--title>EventTarget.addEventListener: capture argument omitted</title-->
<!--link rel="help" href="https://dom.spec.whatwg.org/#dom-eventtarget-addeventlistener"-->
<!--link rel="help" href="https://dom.spec.whatwg.org/#concept-event-dispatch"-->
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #table {
    display: none;
  }
  .col {
    display: inline-block;
  }
  .cell {
    border: 1px solid black;
    padding: 0;
    margin: 1px;
  }
</style>
<div id=log></div>
<div id="table">
  <div id="table-body">
    <div class="col" id="parent">
      <p class="cell">Shady Grove</p>
      <p class="cell" id="target">Over the river, Charlie</p>
    </div>
    <div class="col">
      <p class="cell">Aeolian</p>
      <p class="cell">Dorian</p>
    </div>
  </div>
</div>
<script>
test(function() {
    var event_type = "foo";
    var target = document.getElementById("target");
    var targets = [
        target,
        document.getElementById("parent"),
        document.getElementById("table-body"),
        document.getElementById("table"),
        document.body,
        document.documentElement,
        document,
        window
    ];
    var phases = [
        Event.AT_TARGET,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE
    ];

    var actual_targets = [], actual_phases = [];
    var test_event = function(evt) {
        actual_targets.push(evt.currentTarget);
        actual_phases.push(evt.eventPhase);
    }

    for (var i = 0; i < targets.length; i++) {
        targets[i].addEventListener(event_type, test_event);
    }

    // EDIT
    // var evt = document.createEvent("Event");
    // evt.initEvent(event_type, true, true);
    var evt = new Event(event_type, {"bubbles":true, "cancelable":true});

    target.dispatchEvent(evt);

    for (var i = 0; i < targets.length; i++) {
        targets[i].removeEventListener(event_type, test_event);
    }

    target.dispatchEvent(evt);

    assert_array_equals(actual_targets, targets, "targets");
    assert_array_equals(actual_phases, phases, "phases");
}, "EventTarget.addEventListener with the capture argument omitted");
</script>
